refactor: 레거시 컨트롤러 어노테이션 기반 MVC로 변경
